# data-collector/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 数据采集服务API文档，提供股票、财务、新闻和政策数据采集功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Data Collector API
  version: "1.0"
paths:
...
  /api/v1/stocks/quotes/by-date:
    get:
      consumes:
      - application/json
      description: 获取指定日期所有股票的行情数据
      parameters:
      - description: 交易日期，格式：2006-01-02
        in: query
        name: date
        required: true
        type: string
      - description: 限制返回数量，默认100
        in: query
        name: limit
        type: integer
      - description: 偏移量，默认0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取指定日期的行情数据
      tags:
      - 股票行情
  /api/v1/stocks/quotes/by-symbol:
    get:
      consumes:
      - application/json
      description: 获取指定股票在指定时间范围内的行情数据
      parameters:
      - description: 股票代码，如：000001
        in: query
        name: symbol
        required: true
        type: string
      - description: 开始日期，格式：2006-01-02
        in: query
        name: start_date
        type: string
      - description: 结束日期，格式：2006-01-02
        in: query
        name: end_date
        type: string
      - description: 限制返回数量，默认100
        in: query
        name: limit
        type: integer
      - description: 偏移量，默认0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取指定股票的行情数据
      tags:
      - 股票行情
...
  /collect/stock/quotes:
    post:
      consumes:
      - application/json
      description: 采集指定日期的股票行情数据，支持指定股票代码列表
      parameters:
      - description: 交易日期，格式：2006-01-02
        in: query
        name: date
        required: true
        type: string
      - description: 股票代码列表，用逗号分隔，如：000001.SZ,000002.SZ
        in: query
        name: symbols
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 采集指定日期的股票行情数据
      tags:
      - 股票行情采集
  /collect/stock/quotes/latest:
    post:
      consumes:
      - application/json
      description: 采集最新交易日的股票行情数据，支持指定股票代码列表
      parameters:
      - description: 股票代码列表，用逗号分隔，如：000001.SZ,000002.SZ
        in: query
        name: symbols
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 采集最新的股票行情数据
      tags:
      - 股票行情采集
  /collect/stock/quotes/range:
    post:
      consumes:
      - application/json
      description: 采集指定时间范围的股票行情数据，支持指定股票代码列表
      parameters:
      - description: 开始日期，格式：2006-01-02
        in: query
        name: start_date
        required: true
        type: string
      - description: 结束日期，格式：2006-01-02
        in: query
        name: end_date
        required: true
        type: string
      - description: 股票代码列表，用逗号分隔，如：000001.SZ,000002.SZ
        in: query
        name: symbols
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 采集指定时间范围的股票行情数据
      tags:
      - 股票行情采集
...
schemes:
- http
- https
swagger: "2.0"